+++ src/app/Controllers/Authenticated/Task/DelayTaskController.py
@@ -19,11 +19,12 @@ task_service = TaskService()
 class DelayTask(RequestValidationController):
 
     request_dto = api.model(
-        "Delay Task Request", {
+        "Delay Task Request",
+        {
             "task_id": fields.Integer(required=True),
             "delay_for": fields.Integer(required=True),
             "reason": fields.String(required=True),
-        }
+        },
     )
 
     @requires_jwt
@@ -81,7 +82,6 @@ class DelayTask(RequestValidationController):
 
 @api.route("/<int:task_id>")
 class GetDelayTask(RequestValidationController):
-
     class NullableDateTime(fields.DateTime):
         __schema_type__ = ["datetime", "null"]
         __schema_example__ = "None|2019-09-17T19:08:00+10:00"