+++ src/app/Controllers/Authenticated/Task/TaskActivityController.py
@@ -17,7 +17,7 @@ class TaskActivity(RequestValidationController):
 
     activity_dto = api.model(
         "Activity",
-        {"activity": fields.String(), "activity_timestamp": fields.String(), "event_friendly": fields.String()}
+        {"activity": fields.String(), "activity_timestamp": fields.String(), "event_friendly": fields.String()},
     )
 
     response_dto = api.model("Activity Model", {"activity": fields.List(fields.Nested(activity_dto))})