+++ src/CHANGELOG.md
@@ -16,3 +16,5 @@ A list of important changes that may break things
 11. `PUT` `/task/{id}` returns `204` instead of `200`
 12. `POST` `/task/{id}` returns `204` instead of `200` (create and schedule)
 13. Create, Update and Delete `/task-labels` now return `204`
+14. Create, Update and Delete `/task-type` returns `204`
+15. Transition task returns 204